Implemented seasonal + narrative
Implemented seasonal and narrative endpoints.
Some datetime tokens within the JSON data would be presented as a string, causing the JSON converter to fail parsing it when it was only expecting an integer. Now it checks if the token is a string or integer and reads accordingly.
Previous statistics endpoint is now no longer working (or it is but has changed to different request type). Nevertheless, Rainbow 6 website no longer sends a request to the URL so it is marked obselete.