Improved Example program + separation of data response
Example program now fully includes uses for all of the API that has been implemented.
Previously, DateTime parsing was only done with one format, which did not include the seasonal format given by Ubisoft. This is implemented with the new SeasonalDataResponse class.
DataResponse is now split and broken down to their respective classes (current, seasonal and narrative) to reflect the endpoints that are parsing it.